gateway/requirements.txt
flask
paho-mqtt
grpcio
grpcio-tools
websockets
